+++ cookbook/gravmag_harvester_tensor2.py
@@ -50,7 +50,7 @@ seeds = gm.harvester.sow(
      (3701, 1701, 501, {'density':1000}),
      (3201, 1701, 501, {'density':1000}),
      (2951, 3951, 301, {'density':800}),
-     (2951, 3951, 701, {'density':800})], 
+     (2951, 3951, 701, {'density':800})],
     mesh)
 # Run the inversion and collect the results
 estimate, predicted = gm.harvester.harvest(data, seeds, mesh,
@@ -75,13 +75,13 @@ for i in xrange(len(tensor)):
 for i in xrange(len(tensor)):
     mpl.subplot(2, 2, i + 3)
     residuals = tensor[i] - predicted[i]
-    mpl.title('residuals stddev = %.2f' % (residuals.std())) 
+    mpl.title('residuals stddev = %.2f' % (residuals.std()))
     mpl.hist(residuals, bins=10)
     mpl.xlabel('Residual (Eotvos)')
 mpl.show()
 myv.figure()
 myv.prisms(model, 'density', style='wireframe')
-myv.prisms([mesh[s.i] for s in seeds], 'density')
+myv.prisms(seeds, 'density')
 myv.axes(myv.outline(bounds), ranges=[i*0.001 for i in bounds], fmt='%.1f',
     nlabels=6)
 myv.wall_bottom(bounds)